# go/cmd/policy-enforcer/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DYNAMOS Policy Enforcer API
  description: |
    API for policy enforcement in the DYNAMOS system.
    
    This API provides two levels of abstraction:
    
    ## Policy Enforcer API (`/policy-enforcer/*`)
    High-level, reasoner-agnostic interface for policy validation. This API is designed
    to work with different reasoning engines (eFLINT, Symboleo, JSON-based, etc.) through
    a common interface. Use these endpoints for:
    - **Allowed Clauses**: Query what request types, datasets, archetypes, and compute providers
      are allowed for a specific requester at an organization
    - **Request Validation**: Check if a specific data request is permitted by the policy
    
    ## eFLINT API (`/eflint/*`)
    Low-level interface for direct eFLINT server management. Use these endpoints for:
    - **Instance Management**: Start, stop, and monitor eFLINT server instances
    - **Commands**: Send raw commands to the eFLINT server
    - **State Management (POC)**: Export, import, and checkpoint eFLINT execution graphs
    
    ## Architecture
    ```
    ┌─────────────────────────────────────────────────────────────────────────┐
    │                           HTTP/gRPC Layer                                │
    ├─────────────────────────────────────────────────────────────────────────┤
    │  Policy Enforcer API              │  eFLINT API                         │
    │  /policy-enforcer/*               │  /eflint/*                          │
...
servers:
...
paths:
...
  /policy-enforcer/allowed-request-types:
    get:
      summary: Get allowed request types
      description: |
        Returns all request types that are allowed for a requester at an organization.
        This is a reasoner-agnostic endpoint that works with any configured reasoning engine.
      operationId: getAllowedRequestTypes
      tags:
        - Policy Enforcer
      parameters:
        - $ref: '#/components/parameters/OrganizationParam'
        - $ref: '#/components/parameters/RequesterParam'
      responses:
        '200':
          description: Allowed request types retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllowedClausesResponse'
        '400':
          description: Bad request - missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Reasoner is not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /policy-enforcer/allowed-data-sets:
    get:
      summary: Get allowed data sets
      description: |
        Returns all datasets that are allowed for a requester at an organization.
        This is a reasoner-agnostic endpoint that works with any configured reasoning engine.
      operationId: getAllowedDataSets
      tags:
        - Policy Enforcer
      parameters:
        - $ref: '#/components/parameters/OrganizationParam'
        - $ref: '#/components/parameters/RequesterParam'
      responses:
        '200':
          description: Allowed data sets retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllowedClausesResponse'
        '400':
          description: Bad request - missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Reasoner is not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /policy-enforcer/allowed-archetypes:
    get:
      summary: Get allowed archetypes
      description: |
        Returns all archetypes that are allowed for a requester at an organization.
        This is a reasoner-agnostic endpoint that works with any configured reasoning engine.
      operationId: getAllowedArchetypes
      tags:
        - Policy Enforcer
      parameters:
        - $ref: '#/components/parameters/OrganizationParam'
        - $ref: '#/components/parameters/RequesterParam'
      responses:
        '200':
          description: Allowed archetypes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllowedClausesResponse'
        '400':
          description: Bad request - missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Reasoner is not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /policy-enforcer/allowed-compute-providers:
    get:
      summary: Get allowed compute providers
      description: |
        Returns all compute providers that are allowed for a requester at an organization.
        This is a reasoner-agnostic endpoint that works with any configured reasoning engine.
      operationId: getAllowedComputeProviders
      tags:
        - Policy Enforcer
      parameters:
        - $ref: '#/components/parameters/OrganizationParam'
        - $ref: '#/components/parameters/RequesterParam'
      responses:
        '200':
          description: Allowed compute providers retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllowedClausesResponse'
        '400':
          description: Bad request - missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Reasoner is not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /policy-enforcer/allowed-clauses:
    get:
      summary: Get all allowed clauses
      description: |
        Returns all allowed clauses (request types, data sets, archetypes, and compute providers)
        for a requester at an organization in a single request.
        This is a convenience endpoint that combines all individual allowed-* endpoints.
      operationId: getAllAllowedClauses
      tags:
        - Policy Enforcer
      parameters:
        - $ref: '#/components/parameters/OrganizationParam'
        - $ref: '#/components/parameters/RequesterParam'
      responses:
        '200':
          description: All allowed clauses retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllAllowedClausesResponse'
        '400':
          description: Bad request - missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Reasoner is not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /policy-enforcer/validate:
    post:
      summary: Validate request
      description: |
        Validates whether a specific data request is allowed according to the policy.
        
        This endpoint checks all aspects of the request:
        - Request type is allowed
        - Data set is allowed
        - Archetype is allowed
        - Compute provider is allowed
        - The combination is permitted by the agreement
        
        This is the main endpoint for policy enforcement in DYNAMOS.
      operationId: validateRequest
      tags:
        - Policy Enforcer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateRequestParams'
            examples:
              allowed_request:
                summary: Request that should be allowed
                value:
                  organization: "VU"
                  requester: "jorrit.stutterheim@cloudnation.nl"
                  request_type: "sqlDataRequest"
                  data_set: "wageGap"
                  archetype: "dataThroughTtp"
                  compute_provider: "SURF"
              denied_request:
                summary: Request that should be denied (computeToData was revoked)
                value:
                  organization: "VU"
                  requester: "jorrit.stutterheim@cloudnation.nl"
                  request_type: "sqlDataRequest"
                  data_set: "wageGap"
                  archetype: "computeToData"
                  compute_provider: "SURF"
      responses:
        '200':
          description: Validation completed (check 'allowed' field for result)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Reasoner is not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /policy-enforcer/available-archetypes:
    get:
      summary: Get available archetypes for an organization
      description: |
        Returns all archetypes that are available at an organization level (not requester-specific).
        These are archetypes that the organization has configured as options,
        regardless of which requesters have been authorized to use them.
      operationId: getAvailableArchetypes
      tags:
        - Policy Enforcer
      parameters:
        - $ref: '#/components/parameters/OrganizationOnlyParam'
      responses:
        '200':
          description: Available archetypes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailableValuesResponse'
        '400':
          description: Bad request - missing organization parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Reasoner is not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /policy-enforcer/available-compute-providers:
    get:
      summary: Get available compute providers for an organization
      description: |
        Returns all compute providers that are available at an organization level (not requester-specific).
        These are providers that the organization has configured as options,
        regardless of which requesters have been authorized to use them.
      operationId: getAvailableComputeProviders
      tags:
        - Policy Enforcer
      parameters:
        - $ref: '#/components/parameters/OrganizationOnlyParam'
      responses:
        '200':
          description: Available compute providers retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailableComputeProvidersResponse'
        '400':
          description: Bad request - missing organization parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Reasoner is not running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

# -----------------------------------------------------------------------------
# Components
# -----------------------------------------------------------------------------
components:
...
  parameters:
    OrganizationParam:
      name: organization
      in: query
      required: true
      description: The organization/steward identifier
      schema:
        type: string
      example: VU
    RequesterParam:
      name: requester
      in: query
      required: true
      description: The requester/user identifier (typically an email)
      schema:
        type: string
      example: jorrit.stutterheim@cloudnation.nl
    OrganizationOnlyParam:
      name: organization
      in: query
      required: true
      description: The organization identifier
      schema:
        type: string
      example: VU

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    AllowedClausesResponse:
      type: object
      properties:
        organization:
          type: string
          description: The organization/steward identifier
          example: VU
        requester:
          type: string
          description: The requester/user identifier
          example: jorrit.stutterheim@cloudnation.nl
        values:
          type: array
          items:
            type: string
          description: List of allowed values
          example: ["sqlDataRequest", "genericRequest"]
    AllAllowedClausesResponse:
      type: object
      properties:
        organization:
          type: string
          description: The organization/steward identifier
          example: VU
        requester:
          type: string
          description: The requester/user identifier
          example: jorrit.stutterheim@cloudnation.nl
        request_types:
          type: array
          items:
            type: string
          description: List of allowed request types
          example: ["sqlDataRequest", "genericRequest"]
        data_sets:
          type: array
          items:
            type: string
          description: List of allowed datasets
          example: ["wageGap"]
        archetypes:
          type: array
          items:
            type: string
          description: List of allowed archetypes
          example: ["dataThroughTtp", "reproducableScience"]
        compute_providers:
          type: array
          items:
            type: string
          description: List of allowed compute providers
          example: ["SURF"]
    ValidateRequestParams:
      type: object
      required:
        - organization
        - requester
        - request_type
        - data_set
        - archetype
        - compute_provider
      properties:
        organization:
          type: string
          description: The data steward organization
          example: "VU"
        requester:
          type: string
          description: The user making the request
          example: "jorrit.stutterheim@cloudnation.nl"
        request_type:
          type: string
          description: Type of request
          example: "sqlDataRequest"
        data_set:
          type: string
          description: The dataset being requested
          example: "wageGap"
        archetype:
          type: string
          description: The processing archetype
          example: "dataThroughTtp"
        compute_provider:
          type: string
          description: Where the computation runs
          example: "SURF"
    ValidationResponse:
      type: object
      properties:
        allowed:
          type: boolean
          description: Whether the request is permitted
          example: true
        reason:
          type: string
          description: Explanation for the decision
          example: "Request is permitted by the agreement"
        organization:
          type: string
          description: The organization checked
          example: "VU"
        requester:
          type: string
          description: The requester checked
          example: "jorrit.stutterheim@cloudnation.nl"
        request_type:
          type: string
          description: The request type checked
          example: "sqlDataRequest"
        data_set:
          type: string
          description: The dataset checked
          example: "wageGap"
        archetype:
          type: string
          description: The archetype checked
          example: "dataThroughTtp"
        compute_provider:
          type: string
          description: The compute provider checked
          example: "SURF"
    AvailableValuesResponse:
      type: object
      properties:
        organization:
          type: string
          description: The organization identifier
          example: "VU"
        archetypes:
          type: array
          items:
            type: string
          description: List of available archetypes
          example: ["computeToData", "dataThroughTtp", "reproducableScience"]
    AvailableComputeProvidersResponse:
      type: object
      properties:
        organization:
          type: string
          description: The organization identifier
          example: "VU"
        compute_providers:
          type: array
          items:
            type: string
          description: List of available compute providers
          example: ["SURF", "otherCompany"]

    # -------------------------------------------------------------------------
    # State Management Schemas
    # -------------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "no instance running"
...